splitting params into a new file
